fix: time representation (#8)
* fix: harmonising the time to epoch milliseconds

* refactor: make clientDb private in MainActivity

* fix: add event screen page

* fix: replace individual states to use the entity instead

* refactor: moved viewModel into composable with viewModel function

* refactor: changed comments and types of some classes

+++ app/src/main/java/io/github/drp08/studypal/domain/entities/SessionEntity.kt
@@ -10,8 +10,8 @@ data class SessionEntity(
     @PrimaryKey
     val sessionId: Int,
     val topic: String, // The foreign to the topic name
-    val startTime: Int, // Second of the day
-    val endTime: Int, // Second of the day
+    val startTime: Long, // Epoch millis
+    val endTime: Long, // Epoch millis
     val totalSessions: Int
 ) {
     fun toSerializable(): Session {
