+++ shared/src/main/java/io/github/drp08/studypal/domain/models/Session.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class Session(
     val sessionId: Int,
     val topic: String,
-    val startTime: Int, // Second of the day
-    val endTime: Int, // Second of the day
+    val startTime: Long, // Epoch millis
+    val endTime: Long, // Epoch millis
     val totalSessions: Int
 )
